♻️  refactor : 외부 클릭시 팝업 닫히는 함수 hook으로 분리
