+++ src/@types/type.ts
@@ -1,0 +1,4 @@
+export type useOutSideClickProps = {
+  ref: React.RefObject<HTMLDivElement>;
+  callback: () => void;
+};
